requirements.txt
python-dotenv
openai
langchain
neo4j
networkx
matplotlib
numpy
PyPDF2
python-docx
